fix type, rename test file
fix type
rename test file
add  type
