+++ src/indexer/mod.rs
@@ -7,7 +7,7 @@ mod log_merge_policy;
 mod merge_operation;
 pub mod merge_policy;
 pub mod merger;
-mod merger_sorted_index;
+mod merger_sorted_index_test;
 pub mod operation;
 mod prepared_commit;
 mod segment_entry;
